Add combined_averaging_ssd_avx512()
Speed comparing with original avx2:
 8x8 :  3.61x
16x16:  7.67x
32x32: 14.34x
64x64: 18.31x

Update eb_av1_txb_init_levels_avx2()

Speed comparison:
 4x 4: 2.12x
 4x 8: 2.01x
 4x16: 1.62x
 8x 4: 1.82x
 8x 8: 1.53x
 8x16: 1.47x
 8x32: 1.43x
16x 4: 1.38x
16x 8: 1.49x
16x16: 1.47x
16x32: 1.36x
32x 8: 1.52x
32x16: 1.15x
32x32: 1.24x

Add eb_av1_txb_init_levels_avx512()

About 10% to 30% faster for 16x and 32x.

Fix build errors on branch avx512-average-txb

Temporally activate AVX512 code

Fix a gcc build error on branch avx512-average-txb

Revert "Temporally activate AVX512 code"

This reverts commit 5faee6d3c1a355fe480dcd0a144cff5f25c3e589.

Restore the combined_averaging_ssd_func_ptr_array[] table

+++ Source/Lib/Common/ASM_SSE2/synonyms.h
@@ -51,18 +51,18 @@ static INLINE void _mm_storeh_epi64(__m128i *const p, const __m128i x) {
 }
 
 static INLINE __m128i load8bit_8x2_sse2(const void *const src,
-    const uint32_t strideInByte) {
+    const ptrdiff_t strideInByte) {
     const __m128i s = _mm_loadl_epi64((__m128i *)src);
     return _mm_loadh_epi64((__m128i *)((uint8_t *)src + strideInByte), s);
 }
 
 static INLINE __m128i load_u8_8x2_sse2(const uint8_t *const src,
-    const uint32_t stride) {
+    const ptrdiff_t stride) {
     return load8bit_8x2_sse2(src, sizeof(*src) * stride);
 }
 
 static INLINE __m128i load_u16_4x2_sse2(const uint16_t *const src,
-    const uint32_t stride) {
+    const ptrdiff_t stride) {
     return load8bit_8x2_sse2(src, sizeof(*src) * stride);
 }
